requirements.txt
django==*
mysqlclient==*
djangorestframework==*
django-cors-headers==*
dj-database-url==*
gunicorn==*
psycopg2==*
